+++ src/test/java/org/seedstack/kafka/streams/KafkaStreamsIT.java
@@ -1,5 +1,5 @@
-/**
- * Copyright (c) 2013-2016, The SeedStack authors <http://seedstack.org>
+/*
+ * Copyright © 2013-2019, The SeedStack authors <http://seedstack.org>
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -14,27 +14,29 @@ import javax.inject.Inject;
 import javax.inject.Named;
 import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerRecord;
-import org.fest.assertions.Assertions;
-import org.fest.assertions.Fail;
+import org.assertj.core.api.Assertions;
 import org.junit.Test;
-import org.seedstack.seed.it.AbstractSeedIT;
+import org.junit.runner.RunWith;
+import org.seedstack.seed.testing.junit4.SeedITRunner;
 
-public class KafkaStreamsIT extends AbstractSeedIT {
+@RunWith(SeedITRunner.class)
+public class KafkaStreamsIT {
     private static final String TOPIC = "testStreams";
     static CountDownLatch count = new CountDownLatch(2);
     @Inject
-    @Named("kafkaProducerTest")
+    @Named("producerTest")
     private Producer<Integer, String> producer;
 
     @Test
-    public void streamTest() {
+    public void streamTest() throws InterruptedException {
+        producer.beginTransaction();
         try {
             producer.send(new ProducerRecord<>(TOPIC, 3, "test3"));
             producer.send(new ProducerRecord<>(TOPIC, 4, "test4"));
-            Assertions.assertThat(count.await(20, TimeUnit.SECONDS)).isTrue();
-            producer.close();
+            producer.commitTransaction();
         } catch (Exception e) {
-            Fail.fail(e.getMessage(), e);
+            producer.abortTransaction();
         }
+        Assertions.assertThat(count.await(20, TimeUnit.SECONDS)).isTrue();
     }
 }