+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+# Version 2.0.0 (2019-03-08)
+
+* [new] Automatic retry of failing pollers and streams.
+* [fix] Fix stream and consumer lifecycle.
+* [brk] Reworked API.
+
 # Version 1.1.0 (2018-08-06)
 
-* [chg] Using StreamBuilder instead of the Deprecated KStreamBuilder
+* [chg] Using StreamBuilder instead of the Deprecated KStreamBuilder.
 * [fix] Fix Kafka producer warning when the `client.id` property is defined.
 
 # Version 1.0.0 (2017-11-29)